Create an interface for ChatRepository
