+++ app/src/test/java/com/swent/suddenbump/model/user/UserViewModelTest.kt
@@ -1,7 +1,6 @@
 package com.swent.suddenbump.model.user
 
 import android.location.Location
-import com.swent.suddenbump.model.chat.ChatRepository
 import org.hamcrest.CoreMatchers.`is`
 import org.hamcrest.MatcherAssert.assertThat
 import org.junit.Before
